restructured packages for orders kafka example
