+++ src/main/java/com/ibm/hello/model/kafka/orders/OrderEntity.java
@@ -1,5 +1,4 @@
-package com.ibm.hello.model;
-
+package com.ibm.hello.model.kafka.orders;
 
 public class OrderEntity {
 	public static final String PENDING_STATUS = "pending";
