+++ src/main/java/com/ibm/hello/model/kafka/orders/OrderEvent.java
@@ -1,4 +1,4 @@
-package com.ibm.hello.model;
+package com.ibm.hello.model.kafka.orders;
 
 public class OrderEvent {
